modified the output of the headtracker from int to float in order to get more accuracy

+++ sdk/src/main/java/com/core/realwear/sdk/HFHeadtrackerListener.java
@@ -31,7 +31,6 @@ public interface HFHeadtrackerListener {
     // Head Movement Event
     //
     /////////////////////////////////////////////////////////////////////////////
-    void onHeadMoved(int deltaX, int deltaY);
-
+    void onHeadMoved(float deltaX, float deltaY);
 
 }
